new game now goes to sprite in scrolling background (controlled by arrow keys)
Deleted some unnecessary scenes and scripts. took out gravity from the sprite because if there was gravity the sprite would fall endlessly and you wouldn't be able to scroll. esc doesn't work in the parallax world for some reason, need to fix that. Need to add a collision body in the world so the player can fall onto it